requirements.txt
kafka-python
s3fs
FlightRadar24 